test(server): keep recovery tests in legacy inventory
