+++ crates/harness-server/src/intake/github_coverage_recovery_tests/atomic.rs
@@ -3,10 +3,80 @@ use harness_core::config::isolation::{
     IsolationAvailability, IsolationConfig, IsolationRule, IsolationTier, IsolationTierStatus,
     IsolationTrustClass,
 };
+use harness_workflow::issue_lifecycle::IssueWorkflowInstance;
+use harness_workflow::project_lifecycle::{ProjectWorkflowInstance, ProjectWorkflowStore};
 use harness_workflow::runtime::{
     CommandDispatchOutcome, RuntimeCommandDispatcher, RuntimeKind, RuntimeProfile,
     WorkflowCommandStatus,
 };
+
+#[tokio::test]
+async fn canonical_intake_reuses_legacy_issue_and_project_workflows() -> anyhow::Result<()> {
+    let Some(database_url) = crate::test_helpers::configured_test_database_url()? else {
+        return Ok(());
+    };
+    let dir = crate::test_helpers::tempdir_in_home("harness-test-legacy-intake-")?;
+    let issue_store = IssueWorkflowStore::open_with_database_url(
+        &dir.path().join("issue-workflows"),
+        Some(&database_url),
+    )
+    .await?;
+    let project_store = ProjectWorkflowStore::open_with_database_url(
+        &dir.path().join("project-workflows"),
+        Some(&database_url),
+    )
+    .await?;
+    let project_root = dir.path().join("project");
+    std::fs::create_dir(&project_root)?;
+    let project_id = project_root.to_string_lossy().into_owned();
+    let issue_number = 1708;
+    let mut legacy_issue = IssueWorkflowInstance::new(
+        project_id.clone(),
+        Some("owner/repo".to_string()),
+        issue_number,
+    );
+    legacy_issue.id = workflow_id(&project_id, Some("Owner/Repo"), issue_number);
+    legacy_issue.repo = Some("Owner/Repo".to_string());
+    issue_store.upsert(&legacy_issue).await?;
+    let mut legacy_project =
+        ProjectWorkflowInstance::new(project_id.clone(), Some("owner/repo".to_string()));
+    legacy_project.id =
+        harness_workflow::project_lifecycle::workflow_id(&project_id, Some("Owner/Repo"));
+    legacy_project.repo = Some("Owner/Repo".to_string());
+    project_store.upsert(&legacy_project).await?;
+
+    let coverage = check_github_issue_coverage(
+        Some(&issue_store),
+        None,
+        &project_root,
+        &project_id,
+        REPO,
+        issue_number,
+        IsolationTrustClass::Trusted,
+        None,
+    )
+    .await?;
+    assert!(matches!(
+        coverage,
+        GitHubIssueCoverage::Covered {
+            source: "issue_workflow",
+            ..
+        }
+    ));
+    let idle = project_store.record_idle(&project_id, Some(REPO)).await?;
+
+    assert_eq!(idle.id, legacy_project.id);
+    assert_eq!(project_store.row_count().await?, 1);
+    assert_eq!(issue_store.row_count().await?, 1);
+    assert_eq!(
+        issue_store
+            .get_by_issue(&project_id, Some(REPO), issue_number)
+            .await?
+            .map(|workflow| workflow.id),
+        Some(legacy_issue.id)
+    );
+    Ok(())
+}
 
 #[tokio::test]
 async fn cancelled_recovered_quality_gate_is_reactivated_after_terminal_job() -> anyhow::Result<()>
